+++ packages/storybook/examples/Form.TextInput/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
 
 // For props table
 import TextInput from '@ichef/gypcrete-form/src/TextInput';
@@ -7,6 +8,8 @@ import TextInput from '@ichef/gypcrete-form/src/TextInput';
 import BasicUsage from './BasicUsage';
 
 storiesOf('[Form] TextInput', module)
-    .addWithInfo('basic usage', BasicUsage)
+    .add('basic usage',
+        withInfo()(BasicUsage)
+    )
     // Props table
     .addPropsTable(() => <TextInput />);