+++ packages/storybook/examples/Tag/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
 
 // For props table
 import Tag from '@ichef/gypcrete/src/Tag';
@@ -8,7 +9,11 @@ import BasicTagExample from './BasicTag';
 import TagWithParentColorExample from './TagWithParentColor';
 
 storiesOf('Tag', module)
-    .addWithInfo('basic usage', BasicTagExample)
-    .addWithInfo('with parent color', TagWithParentColorExample)
+    .add('basic usage',
+        withInfo()(BasicTagExample)
+    )
+    .add('with parent color',
+        withInfo()(TagWithParentColorExample)
+    )
     // Props table
     .addPropsTable(() => <Tag />);