+++ packages/storybook/examples/Text/Ellipsis.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Text from '@ichef/gypcrete/src/Text';
 import TextEllipsis from '@ichef/gypcrete/src/TextEllipsis';
-import DebugBox from '../DebugBox';
+import DebugBox from 'utils/DebugBox';
 
 const LONG_LABEL =
     `Lorem ipsum dolor sit amet, consectetur adipiscing elit.