+++ packages/storybook/examples/Text/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
 
 // For props table
 import Text, { PureText } from '@ichef/gypcrete/src/Text';
@@ -9,8 +10,14 @@ import EllipsisExample from './Ellipsis';
 import TextWithStatusIconExample from './TextWithStatusIcon';
 
 storiesOf('Text', module)
-    .addWithInfo('basic usage', BasicTextExample)
-    .addWithInfo('ellipsis-cropped', EllipsisExample)
-    .addWithInfo('with statusIcon', TextWithStatusIconExample)
+    .add('basic usage',
+        withInfo()(BasicTextExample)
+    )
+    .add('ellipsis-cropped',
+        withInfo()(EllipsisExample)
+    )
+    .add('with statusIcon',
+        withInfo()(TextWithStatusIconExample)
+    )
     // Props table
     .addPropsTable(() => <Text />, [PureText]);