+++ packages/storybook/examples/TextLabel/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
 
 // For props table
 import TextLabel, { PureTextLabel } from '@ichef/gypcrete/src/TextLabel';
@@ -10,9 +11,17 @@ import TextLabelWithTextExample from './TextLabelWithText';
 import Editable from './Editable';
 
 storiesOf('TextLabel', module)
-    .addWithInfo('basic usage', BasicTextLabelExample)
-    .addWithInfo('with status', TextLabelWithStatusExample)
-    .addWithInfo('<Text> in child', TextLabelWithTextExample)
-    .addWithInfo('Editable', Editable)
+    .add('basic usage',
+        withInfo()(BasicTextLabelExample)
+    )
+    .add('with status',
+        withInfo()(TextLabelWithStatusExample)
+    )
+    .add('<Text> in child',
+        withInfo()(TextLabelWithTextExample)
+    )
+    .add('Editable',
+        withInfo()(Editable)
+    )
     // Props table
     .addPropsTable(() => <TextLabel />, [PureTextLabel]);