+++ packages/storybook/examples/Tooltip/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
 
 // For props table
 import Tooltip from '@ichef/gypcrete/src/Tooltip';
@@ -9,8 +10,14 @@ import TooltipBottomPlacementExample from './TooltipBottomPlacement';
 import TooltipCustomArrowStyleExample from './TooltipCustomArrowStyle';
 
 storiesOf('Tooltip', module)
-    .addWithInfo('basic usage', BasicTooltipExample)
-    .addWithInfo('bottom placement', TooltipBottomPlacementExample)
-    .addWithInfo('custom arrow style', TooltipCustomArrowStyleExample)
+    .add('basic usage',
+        withInfo()(BasicTooltipExample)
+    )
+    .add('bottom placement',
+        withInfo()(TooltipBottomPlacementExample)
+    )
+    .add('custom arrow style',
+        withInfo()(TooltipCustomArrowStyleExample)
+    )
     // Props table
     .addPropsTable(() => <Tooltip />);